replace the getOutput function to return vector instead of an array pointer.

+++ protocols/YaoSingleExecution/YaoSEParty.cpp
@@ -2,6 +2,7 @@
 // Created by moriya on 15/2/17.
 //
 #ifndef _WIN32
+
 #include "YaoSEParty.h"
 
 YaoSEParty::YaoSEParty(int id, string circuitFile, string ip, int port, string inputFile)
@@ -121,7 +122,7 @@ int main(int argc, char* argv[]) {
         auto out = party.getOutput();
         cout << "result: " << endl;
         for (int i = 0; i < cf->n3; i++) {
-            cout << out[i] << " ";
+            cout << (int)out[i] << " ";
         }
         cout << endl;
     }
@@ -156,7 +157,7 @@ int main(int argc, char* argv[]) {
         auto out = party.getOutput();
         cout << "result: " << endl;
         for (int i = 0; i < cf->n3; i++) {
-            cout << out[i] << " ";
+            cout << (int)out[i] << " ";
         }
         cout << endl;
     }
